Refactor to break out connecting to an AP into a separate component, available outside of fragments
